refactor, stop overriding console.log
