+++ src/app/communication/sync/sync-renderer-communication.js
@@ -20,7 +20,7 @@
 import messages from '../messages'
 import type { SyncSender } from './sync'
 import type { SyncRendererCommunication } from './sync-communication'
-import type { SerializedLogCaches } from '../../logging/log-cache-bundle'
+import type { SerializedLogCaches } from '../../logger/log-cache-bundle'
 import type { LogDTO } from '../dto'
 import type { RavenData } from '../../bug-reporting/bug-reporter-metrics'
