+++ src/app/mysterion.js
@@ -42,8 +42,8 @@ import IdentityDTO from '../libraries/mysterium-tequilapi/dto/identity'
 import type { CurrentIdentityChangeDTO } from './communication/dto'
 import type { EnvironmentCollector } from './bug-reporting/environment/environment-collector'
 import {BugReporterMetrics, METRICS, TAGS} from '../app/bug-reporting/bug-reporter-metrics'
-import BackendLogBootstrapper from './logging/backend-log-bootstrapper'
-import LogCache from './logging/log-cache'
+import BackendLogBootstrapper from './logger/backend-log-bootstrapper'
+import LogCache from './logger/log-cache'
 import SyncCallbacksInitializer from './sync-callbacks-initializer'
 import SyncReceiverMainCommunication from './communication/sync/sync-main-communication'
 import { SyncIpcReceiver } from './communication/sync/sync-ipc'
