+++ src/renderer/dependencies/modules/bug-reporting.js
@@ -27,7 +27,7 @@ import type { EnvironmentCollector } from '../../../app/bug-reporting/environmen
 import SyncSenderRendererCommunication from '../../../app/communication/sync/sync-renderer-communication'
 import { SyncIpcSender } from '../../../app/communication/sync/sync-ipc'
 import type { SyncRendererCommunication } from '../../../app/communication/sync/sync-communication'
-import FrontendLogBootstrapper from '../../../app/logging/frontend-log-bootstrapper'
+import FrontendLogBootstrapper from '../../../app/logger/frontend-log-bootstrapper'
 import { BugReporterMetrics } from '../../../app/bug-reporting/bug-reporter-metrics'
 
 function bootstrap (container: Container) {
