+++ test/unit/app/bug-reporting/environment/renderer-environment-collector.spec.js
@@ -21,7 +21,7 @@ import { beforeEach, describe, expect, it } from '../../../../helpers/dependenci
 import RendererEnvironmentCollector
   from '../../../../../src/app/bug-reporting/environment/renderer-environment-collector'
 import type { SyncRendererCommunication } from '../../../../../src/app/communication/sync/sync-communication'
-import type { SerializedLogCaches } from '../../../../../src/app/logging/log-cache-bundle'
+import type { SerializedLogCaches } from '../../../../../src/app/logger/log-cache-bundle'
 import type { RavenData } from '../../../../../src/app/bug-reporting/bug-reporter-metrics'
 import { TAGS } from '../../../../../src/app/bug-reporting/bug-reporter-metrics'
